edi: make send failed job retryable
Send jobs might fail due to an external service being not responsive.
If the job is simply failed, a new one will be spawned and might encour in the same error again,
possibly leading to an high number of duplicated jobs for the same record.

Yet, when a RetryableJobError is raised, the job will be set back into pending state
and will be nicely retried based on jobs configuration.